First release on GitHub
- Add build option EMPTY_ATTESTATION_CERT to produce firmware without
  attestation certificate (needs to be initialized later)
- Add vendor command U2F_ATTESTATION_CERT (0x40) to initialize device
  with given attestation certificate
- Assign unique VID:PID

+++ src/u2f-apdu.h
@@ -6,7 +6,7 @@ void
 u2f_apdu_init (void);
 
 int
-u2f_apdu_command_do (uint8_t *apdu, uint8_t len,
+u2f_apdu_command_do (uint8_t *apdu, uint32_t len,
                      uint8_t *resp, uint32_t *resp_len);
 
 #endif
